requirements.txt
arabic-reshaper==2.1.4
asgiref==3.4.1
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
cssselect2==0.7.0
distlib==0.3.5
Django==4.0
django-cors-headers==3.13.0
django-embed-video==1.4.8
django-modeltranslation==0.18.5
django-paypal==2.1
django-rosetta==0.9.8
filelock==3.7.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
joblib==1.2.0
lxml==4.9.1
nltk==3.7
numpy==1.23.4
oscrypto==1.3.0
pandas==1.5.1
Pillow==9.3.0
pipenv==2022.7.4
pisa==3.0.33
platformdirs==2.5.2
polib==1.1.1
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
regex==2022.9.13
reportlab==3.6.12
requests==2.28.1
rosetta==0.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
virtualenv==20.15.1
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
